Add Light objects to list
Made LightController a singleton
Adjusted logic in LightObject to dynamically add lights to light controller list on awake.
This logic will need to be changed when we set up the game modes, but doing it this way shoudl allow us to expend the number of lights in the list easier.